+++ packages/headless/src/controllers/insight/result-list/headless-insight-interactive-result.ts
@@ -1,5 +1,5 @@
 import {pushRecentResult} from '../../../features/recent-results/recent-results-actions';
-import {logDocumentOpen} from '../../../features/result/result-analytics-actions';
+import {logDocumentOpen} from '../../../features/result/result-insight-analytics-actions';
 import {InsightEngine} from '../../../app/insight-engine/insight-engine';
 import {Result} from '../../../api/search/search/result';
 import {